pyproject.toml
[project]
name = "lora-peft-finetuning"
version = "0.1.0"
description = "LoRA fine-tuning with PEFT, Hugging Face Transformers, and wandb logging"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "bert-score>=0.3.13",
    "bitsandbytes>=0.49.2",
    "datasets>=4.8.4",
    "jupyter>=1.1.1",
    "peft>=0.18.1",
    "python-dotenv>=1.2.2",
    "rouge-score>=0.1.2",
    "transformers>=5.5.0",
    "trl>=1.0.0",
    "wandb>=0.19.0",
]

[project.scripts]
train = "train:train"
evaluate = "evaluate:evaluate"
